example file for loadbalancing. updated float32 to int32

+++ edgecast/originv3/api_common.go
@@ -583,7 +583,7 @@ type PatchAdnIdParams struct {
 	MediaType string
 
 	// Customer Origin Id
-	Id float32
+	Id int32
 
 	// The modified Customer Origin
 	CustomerOriginRequest CustomerOriginRequest
